refactor(mercury): split packet.rs by parse/build/legacy seams (#244)
packet.rs was 734 lines (over the 700-line cap). Split into the
existing `packet/` directory (which already held proptest test files):

  packet/mod.rs    — flag constants, ParsedPacket struct + impl,
                     module declarations + pub use re-exports
  packet/parse.rs  — parse_incoming (footer-strip path with pop macros)
  packet/build.rs  — build_outgoing, build_outgoing_fragmented,
                     build_fragmented_bundle, FRAGMENT_BODY_SIZE
  packet/legacy.rs — Packet, PacketFlags (channel/nub stub shim)
  packet/tests.rs  — existing inline unit tests
  packet/{parse_proptest,proptest_round_trip,replay_smoke}.rs unchanged

Public surface preserved at `cimmeria_mercury::packet::*` — external
callers in `crates/mercury/src/{channel,codec}.rs` and the services
crate keep their imports unchanged.

All packet tests pass; clippy clean; cargo fmt applied.
